refactor: moving SourceIterMarker into source_iter_marker.rs
